// src/app/models/product.ts
export class Product{
       id: number;
        brandName: string;
        productName: string;
        quantity: number;
        price: number;
        mrp: number;
        imgUrl: string;
        offerText: string;
}

export class selectedProd {
  id: number;
  quantity: number;
  price: number;
}